
\name{VarComp}
\alias{VarComp}
\title{Variance Components simulated data}
\description{
  Data frame has 30 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{
      a numeric vector giving a response "y".
    }
    \item{block1}{
		a factor identifying a blocking factor (1)
    }
    \item{block2}{
		a factor identifying a blocking factor (2)
    }
   \item{onesubject}{
      a factor identifying the subjects (only one subject).
    }
  }
}
\details{
 This dataset is a simulated set of observations to exemplify
 the use of some methods in the \code{hoalme} package.
 It provides a dataset suitable to be modelled by a variance
 components model, each blocking factor representing a random
 factor.
}

\keyword{datasets}
